pyproject.toml
[tool.poetry]
name = "crowdcomms-django-test"
version = "0.1.0"
description = ""
authors = ["Ben Harling <bharling@crowdcomms.co.uk>"]
readme = "README.md"
packages = [{include = "crowdcomms_django_test"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.14.0"
haversine = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-django = "^4.5.2"
ipython = "^8.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
